Add connection timeouts.
If the connection to FS doesn't end up gracefully, the call to ReadBytes
on the readHeaders method blocks undefinitely waiting for data, meaning
that even though the connection seems to be up, nothing is being
received or sent.

By associating a timeout to the connection, the bufio's ReadBytes
function will throw a timeout error if there's no data for that specific
period of time, or a more specific IO error if the connection ended
abruptly. This way we can check if the error is indeed a timeout, and
carry on with our normal operations, or if the error came from bufio's
ReadBytes function, in which case we would like to catch the error, log
it, and try and reconnect to FS.